# RECUPERACION/DWES/ApiLaravel/ApiLaravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Productos",
        "description": "API de productos",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/productos": {
            "get": {
                "tags": [
                    "productos"
                ],
                "summary": "Obtener lista de productos",
                "description": "Devuelve una lista paginada de productos con su categoría asociada",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "productos"
                ],
                "summary": "Crear un producto",
                "operationId": "48b59f126c849beb4d4de4ff83b7945f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "productos"
                ],
                "summary": "Obtener un producto",
                "description": "Obtener un producto por su id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "productos"
                ],
                "summary": "Actualizar un producto",
                "operationId": "20b20d8e59992eb88abad7dcef3db6d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "4c0c7680428675134afd4d793f934f67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Categoria": {
                "title": "Categoria",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Electrónica"
                    }
                },
                "type": "object"
            },
            "Producto": {
                "title": "Producto",
                "required": [
                    "id",
                    "nombre",
                    "precio",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Ratón inalámbrico"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Ratón con conexión Bluetooth"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "productos",
            "description": "productos"
        }
    ]
}